# mentor-api-timetable/src/common/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "Interactive Mentor API documentation"
  version: "1.0.0"
  title: "Mentor API Doc"
host: "localhost:9200"
basePath: "/"
...
schemes:
  - "http"
  - "https"
paths:
...
  /employees:
    post:
      tags:
        - "Employees"
      summary: "Adding Employee"
      description: ""
      operationId: "addEmployee"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Adding Employee"
          required: true
          schema:
            $ref: "#/definitions/Employee"
      responses:
        "405":
          description: "Invalid input"
    get:
      tags:
        - "Employees"
      summary: "Fetch all Employees"
      description: "Retrieve all Employees"
      operationId: "getEmployees"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              schema:
                $ref: "#/definitions/Employee"
        "404":
          description: "Notfound"
          schema:
            type: "String"
...
  /employees/{id}:
    put:
      tags:
        - "Employees"
      summary: "Modifying an existing exployee"
      description: ""
      operationId: "changeEmployee"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Editing Employee object"
          required: true
          schema:
            $ref: "#/definitions/Employee"
      responses:
        "405":
          description: "Invalid input"
    get:
      tags:
        - "Employees"
      summary: "Fetch Employee by object id"
      description: "Retrieve the Employee"
      operationId: "getEmployee"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "object Id"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "Employee"
        "404":
          description: "Notfound"
          schema:
            type: "String"
    delete:
      tags:
        - "Employees"
      summary: "Remove an employee"
      description: "Delete an employee"
      operationId: "deleteEmployee"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "object Id"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
        "404":
          description: "Notfound"
          schema:
            type: "String"
...
  /employees/enroll/{id}:
    put:
      tags:
        - Employees
      summary: "Modifying an existing exployee"
      description: ""
      operationId: "updateEmployeeEnroll"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Editing Employee object"
          required: true
          schema:
            $ref: "#/definitions/Employee"
      responses:
        "405":
          description: "Invalid input"
  /employees/submit/{id}:
    put:
      tags:
        - Employees
      summary: "Modifying an existing exployee"
      description: ""
      operationId: "updateEmployeeSubmit"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Editing Employee object"
          required: true
          schema:
            $ref: "#/definitions/Employee"
      responses:
        "405":
          description: "Invalid input"
  /employees/student/analytics/{type}:
    get:
      tags:
        - "Employees"
      summary: "Analytics API"
      description: "Retrieve employee analytics"
      operationId: "getEmployeesForAnalytics"
      produces:
        - "application/json"
      parameters:
        - name: "type"
          in: "path"
          description: "Total Employee | Administrative | Faculty | Support Staff | Total Student" 
          required: true
          type: "string"
          enum:
            - "Total Employee"
            - "Administrative"
            - "Faculty"
            - "Support Staff"
            - "Total Student"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "Employee"
        "404":
          description: "Notfound"
          schema:
            type: "String"
  /employees/{instituteId}/teaching-employee:
    get:
      tags:
        - "Employees"
      summary: "Fetch Teaching Employee"
      description: "Retrieve the Teaching Employee"
      operationId: "getTeachingEmployees"
      produces:
        - "application/json"
      parameters:
        - name: "instituteId"
          in: "path"
          description: "object Id"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "Employee"
        "404":
          description: "Notfound"
          schema:
            type: "String"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Employee:
    type: "object"
    properties:
      _id:
        type: "string"
      instituteId:
        type: "string"
      _empId:
        type: "number"
      empId:
        type: "string"
      isActive:
        type: "boolean"
      employeeType:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      contactNumber:
        type: "string"
      emailId:
        type: "string"
      dateOfBirth:
        type: "string"
        format: "date"
      gender:
        type: "string"
      bloodGroup:
        type: "string"
      nationality:
        type: "string"
      religion:
        type: "string"
      casteType:
        type: "string"
      casteName:
        type: "string"
      motherTongue:
        type: "string"
      aadhaarNumber:
        type: "string"
      emergencyContactPersonFirstName:
        type: "string"
      emergencyContactPersonLastName:
        type: "string"
      emergencyContactMobileNumber:
        type: "string"
      permanentAddressLine1:
        type: "string"
      permanentAddressLine2:
        type: "string"
      permanentAddressCountry:
        type: "string"
      permanentAddressState:
        type: "string"
      permanentAddressCity:
        type: "string"
      permanentAddressZipCode:
        type: "string"
      copySameAddress:
        type: "boolean"
      communicationAddressLine1:
        type: "string"
      communicationAddressLine2:
        type: "string"
      communicationAddressCountry:
        type: "string"
      communicationAddressState:
        type: "string"
      communicationAddressCity:
        type: "string"
      communicationAddressZipCode:
        type: "string"
      bankAccountHolderName:
        type: "string"
      bankAccountAccountNumber:
        type: "string"
      bankAccountIFSCCode:
        type: "string"
      panNumber:
        type: "string"
      experiences:
        type: "array"
        items:
          $ref: "#/definitions/EmployeeExperience"
      dependents:
        type: "array"
        items:
          $ref: "#/definitions/EmployeeDependent"
      academicInfos:
        type: "array"
        items:
          $ref: "#/definitions/EmployeeAcademicInfo"
    xml:
      name: "Employee"
  EmployeeAcademicInfo:
    type: "object"
    properties:
      courseType:
        type: "string"
      boardName:
        type: "string"
      schoolCollegeName:
        type: "string"
      overallPercentage:
        type: "string"
      courseName:
        type: "string"
      yearOfPassing:
        type: "string"
    xml:
      name: "EmployeeAcademicInfo"
  EmployeeDependent:
    type: "object"
    properties:
      dependentType:
        type: "string"
      mobileNumber:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      emailId:
        type: "string"
      occupation:
        type: "string"
    xml:
      name: "EmployeeDependent"
  EmployeeExperience:
    type: "object"
    properties:
      employerName:
        type: "string"
      jobTitle:
        type: "string"
      employmentType:
        type: "string"
      location:
        type: "string"
      startDate:
        type: "date"
      endDate:
        type: "date"
    xml:
      name: "EmployeeExperience"
...